third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/gatherElements.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] gatherElements float16 2D input and uint32 indices options.axis=1
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherElements float16 actual -Infinity should be close enough to expected -68.9375 by ULP distance: expected a number less than or equal to 0 but got 10161
[FAIL] gatherElements float16 2D input and int32 indices options.axis=1
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherElements float16 actual -Infinity should be close enough to expected -68.9375 by ULP distance: expected a number less than or equal to 0 but got 10161
[FAIL] gatherElements float16 2D input and int32 indices options.axis=0
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherElements float16 actual Infinity should be close enough to expected -26.15625 by ULP distance: expected a number less than or equal to 0 but got 21130
[FAIL] gatherElements float16 3D input and int32 indices options.axis=0
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherElements float16 actual 0 should be close enough to expected 89.0625 by ULP distance: expected a number less than or equal to 0 but got 21905
[FAIL] gatherElements float16 3D input and int32 negative indices
  assert_less_than_equal: assert_array_approx_equals_ulp: test gatherElements float16 actual 0 should be close enough to expected 89.0625 by ULP distance: expected a number less than or equal to 0 but got 21905
Harness: the test ran to completion.
